requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2020.4.5.2
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
coverage==5.1
distlib==0.3.0
Django==2.2.13
django-discover-runner==1.0
django-prbac==0.0.8
django-role-permissions==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.16
Faker==0.9.1
filelock==3.0.12
flake8==3.8.2
gunicorn
identify==1.4.18
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==1.5.0
Jinja2==2.11.2
jsonfield==2.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.4.0
nodeenv==1.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.4.0
psycopg2-binary==2.8.4
py==1.8.2
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.24.0
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.1.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.21
wcwidth==0.2.4
zipp==3.1.0
